backend/requirements.txt
fastapi
uvicorn
playwright
langchain
langchain-community
langchain-text-splitters
openai
pinecone
python-dotenv
beautifulsoup4 # LangChain's WebBaseLoader dependency
lxml # LangChain's UnstructuredURLLoader dependency
html2text
httpx # For async requests with Playwright
anthropic # For Claude API as alternative to OpenAI
python-magic-bin # For MIME type detection on Windows
pypdf # For PDF parsing
python-docx # For Word document parsing
